+++ game/src/shared/game/Bomb.h
@@ -6,7 +6,7 @@
 #include <vector>
 #include "../Util.h"
 
-#define FUSE_TIME_MILLIS 3000
+#define FUSE_TIME_SEC 3
 
 class Bomb{
 public:
@@ -19,6 +19,7 @@ public:
     [[nodiscard]] bool ShouldExplode() const;
     void animate();
     std::vector<TileOnFire> boom(Map *map);
+    static int CreateBombTimerfd();
 };
 
 #endif //BOOMBERMAN_BOMB_H
